stand-alone useDerivedEntitySearchHits
